Move lists hooks to their own subfolder
